requirements.txt
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
